+++ modules/ec2/install_package.sh
@@ -1,0 +1,8 @@
+#! /bin/bash
+
+# update the machine
+sudo yum update -y
+sleep 5s
+
+# packaged to be installed
+sudo yum install htop -y