+++ modules/ec2/output.tf
@@ -5,3 +5,7 @@ output "PublicIP" {
 output "PrivateIP" {
   value = aws_eip.elastic_ip_addr.private_ip
 }
+
+output "Ec2ID" {
+  value = resource.aws_instance.ec2.id
+}